:memo: fix formater functions
